refactor: move version checker into app shell
