+++ mpvqc/services/__init__.py
@@ -8,10 +8,3 @@ from .file_startup import FileStartupService as FileStartupService
 from .resource import read_input_conf as read_input_conf
 from .resource import read_mpv_conf as read_mpv_conf
 from .state import StateService as StateService
-from .version_checker import HOME_URL as HOME_URL
-from .version_checker import CheckOutcome as CheckOutcome
-from .version_checker import NewVersionAvailable as NewVersionAvailable
-from .version_checker import ServerError as ServerError
-from .version_checker import ServerNotReachable as ServerNotReachable
-from .version_checker import UpToDate as UpToDate
-from .version_checker import VersionCheckerService as VersionCheckerService
