+++ mpvqc/shell/wiring.py
@@ -14,13 +14,14 @@ def bindings(binder: inject.Binder) -> None:
     import inject
     from PySide6.QtCore import QSettings
 
-    from mpvqc.shell.services import QuitService, ShellSettingsService
+    from mpvqc.shell.services import QuitService, ShellSettingsService, VersionCheckerService
 
     def shell_settings_service() -> ShellSettingsService:
         return ShellSettingsService(inject.instance(QSettings))
 
     binder.bind_to_constructor(QuitService, QuitService)
     binder.bind_to_constructor(ShellSettingsService, shell_settings_service)
+    binder.bind_to_constructor(VersionCheckerService, VersionCheckerService)
 
 
 def register_qml_types() -> None:
