+++ pyproject.toml
@@ -255,7 +255,6 @@ files = [
     "mpvqc/services/file_startup.py",
     "mpvqc/services/resource.py",
     "mpvqc/services/state.py",
-    "mpvqc/services/version_checker.py",
     "mpvqc/settings.py",
     "mpvqc/shared.py",
     "mpvqc/shell/__init__.py",
@@ -271,6 +270,7 @@ files = [
     "mpvqc/shell/services/__init__.py",
     "mpvqc/shell/services/quit.py",
     "mpvqc/shell/services/settings.py",
+    "mpvqc/shell/services/version_checker.py",
     "mpvqc/shell/services/vocabulary.py",
     "mpvqc/shell/viewmodels/__init__.py",
     "mpvqc/shell/viewmodels/about.py",
